fix: harden crash breadcrumb parsing
